requirements.txt
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.0
aiohttp==3.9.1
asyncio-mqtt==0.16.1
pydantic==2.5.2
pydantic-settings==2.1.0
loguru==0.7.2
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.13.1
fastapi==0.104.1
uvicorn==0.24.0
